upstream service: move out types and update them
Move out the types from the upstream service module and into a dedicated type file.

Also, remove the branch tree parameter from the resolutions passed in to the integrate call
