+++ apps/desktop/src/lib/upstream/upstreamIntegrationService.ts
@@ -7,119 +7,15 @@ import { isDefined } from '@gitbutler/ui/utils/typeguards';
 import { get } from 'svelte/store';
 import type { BranchStack, PatchSeries } from '$lib/branches/branch';
 import type { Project } from '$lib/project/project';
+import type {
+	BaseBranchResolution,
+	BaseBranchResolutionApproach,
+	BranchStatusesResponse,
+	IntegrationOutcome,
+	Resolution,
+	StackStatusesWithBranches
+} from './types';
 import type { LatestBranchLookupService } from '@gitbutler/shared/branches/latestBranchLookupService';
-
-export type StackStatus = {
-	treeStatus: TreeStatus;
-	branchStatuses: NameAndBranchStatus[];
-};
-
-export type NameAndBranchStatus = {
-	name: string;
-	status: BranchStatus;
-};
-
-export type BranchStatus =
-	| {
-			type: 'empty' | 'integrated' | 'saflyUpdatable';
-	  }
-	| {
-			type: 'conflicted';
-			subject: {
-				rebasable: boolean;
-			};
-	  };
-
-export function stackFullyIntegrated(stackStatus: StackStatus): boolean {
-	return (
-		stackStatus.branchStatuses.every((branchStatus) => branchStatus.status.type === 'integrated') &&
-		stackStatus.treeStatus.type === 'empty'
-	);
-}
-
-export type TreeStatus = {
-	type: 'empty' | 'conflicted' | 'saflyUpdatable';
-};
-
-export type StackStatusInfo = { stack: BranchStack; status: StackStatus };
-
-export type StackStatusesWithBranches =
-	| {
-			type: 'upToDate';
-	  }
-	| {
-			type: 'updatesRequired';
-			subject: StackStatusInfo[];
-	  };
-
-export type ResolutionApproach = {
-	type: 'rebase' | 'merge' | 'unapply' | 'delete';
-};
-
-export type Resolution = {
-	branchId: string;
-	branchTree: string;
-	approach: ResolutionApproach;
-};
-
-export type BaseBranchResolutionApproach = 'rebase' | 'merge' | 'hardReset';
-
-export type BaseBranchResolution = {
-	targetCommitOid: string;
-	approach: { type: BaseBranchResolutionApproach };
-};
-
-export type IntegrationOutcome = {
-	archivedBranches: string[];
-};
-
-export function getBaseBranchResolution(
-	targetCommitOid: string | undefined,
-	approach: BaseBranchResolutionApproach
-): BaseBranchResolution | undefined {
-	if (!targetCommitOid) return;
-
-	return {
-		targetCommitOid,
-		approach: { type: approach }
-	};
-}
-
-export function getResolutionApproach(statusInfo: StackStatusInfo): ResolutionApproach {
-	if (stackFullyIntegrated(statusInfo.status)) {
-		return { type: 'delete' };
-	}
-
-	if (statusInfo.stack.allowRebasing) {
-		return { type: 'rebase' };
-	}
-
-	return { type: 'merge' };
-}
-
-export function sortStatusInfo(a: StackStatusInfo, b: StackStatusInfo): number {
-	if (
-		(!stackFullyIntegrated(a.status) && !stackFullyIntegrated(b.status)) ||
-		(stackFullyIntegrated(a.status) && stackFullyIntegrated(b.status))
-	) {
-		return (a.stack?.name || 'Unknown').localeCompare(b.stack?.name || 'Unknown');
-	}
-
-	if (stackFullyIntegrated(a.status)) {
-		return 1;
-	} else {
-		return -1;
-	}
-}
-
-type BranchStatusesResponse =
-	| {
-			type: 'upToDate';
-	  }
-	| {
-			type: 'updatesRequired';
-			subject: [string, StackStatus][];
-	  };
 
 export class UpstreamIntegrationService {
 	constructor(
